Add Process folder
