+++ cardano-testnet/src/Parsers/Cardano.hs
@@ -12,9 +12,9 @@ import qualified Options.Applicative as OA
 import           Cardano.CLI.Common.Parsers hiding (pNetworkId)
 
 import           Testnet.Cardano
-import           Testnet.Util.Cli
-import           Testnet.Util.Runtime (readNodeLoggingFormat)
-import           Testnet.Utils
+import           Testnet.Process.Cli
+import           Testnet.Property.Utils
+import           Testnet.Runtime (readNodeLoggingFormat)
 
 newtype CardanoOptions = CardanoOptions
   { testnetOptions :: CardanoTestnetOptions
